# internal/openapi/swagger.yaml
basePath: /api
definitions:
...
  api.APIErrorEnvelope:
    properties:
      code:
        example: BAD_REQUEST
        type: string
      error:
        example: true
        type: boolean
      message:
        example: invalid request
        type: string
    type: object
...
  api.DNSRouteSettingsDTO:
    properties:
      autoRefreshEnabled:
        example: true
        type: boolean
      refreshDailyTime:
        example: "03:00"
        type: string
      refreshIntervalHours:
        example: 24
        type: integer
      refreshMode:
        example: interval
        type: string
    type: object
...
  api.DownloadSettingsDTO:
    properties:
      routeTag:
        example: direct
        type: string
    type: object
...
  api.LoggingSettingsDTO:
    properties:
      appMaxEntries:
        example: 5000
        type: integer
      enabled:
        example: true
        type: boolean
      logLevel:
        example: info
        type: string
      maxAge:
        example: 2
        type: integer
      singboxLogLevel:
        enum:
        - trace
        - debug
        - info
        - warn
        - error
        - fatal
        - panic
        example: trace
        type: string
      singboxMaxEntries:
        example: 5000
        type: integer
    type: object
...
  api.OkData:
    properties:
      ok:
        example: true
        type: boolean
    type: object
  api.OkResponse:
    properties:
      data:
        $ref: '#/definitions/api.OkData'
      success:
        example: true
        type: boolean
    type: object
...
  api.PingCheckDefaultsDTO:
    properties:
      deadInterval:
        example: 120
        type: integer
      failThreshold:
        example: 3
        type: integer
      interval:
        example: 30
        type: integer
      method:
        example: http
        type: string
      target:
        example: https://www.google.com
        type: string
    type: object
  api.PingCheckSettingsDTO:
    properties:
      defaults:
        $ref: '#/definitions/api.PingCheckDefaultsDTO'
      enabled:
        example: true
        type: boolean
    type: object
...
  api.ServerSettingsDTO:
    properties:
      interface:
        example: ""
        type: string
      port:
        example: 8080
        type: integer
    type: object
...
  api.SettingsData:
    properties:
      authEnabled:
        example: false
        type: boolean
      disableMemorySaving:
        example: false
        type: boolean
      dnsRoute:
        $ref: '#/definitions/api.DNSRouteSettingsDTO'
      download:
        $ref: '#/definitions/api.DownloadSettingsDTO'
      logging:
        $ref: '#/definitions/api.LoggingSettingsDTO'
      monitoringExcludedTunnels:
        example:
        - tn-1
        - sys-2
        items:
          type: string
        type: array
      pingCheck:
        $ref: '#/definitions/api.PingCheckSettingsDTO'
      schemaVersion:
        example: 16
        type: integer
      server:
        $ref: '#/definitions/api.ServerSettingsDTO'
      updates:
        $ref: '#/definitions/api.UpdateSettingsDTO'
      usageLevel:
        description: |-
          UsageLevel controls which UI sections are visible to the user.
          Filtering is frontend-only — the API does not enforce it.
          enums: expert,advanced,basic
          First enum is the prism mock default (prism picks the first
          enum value over the example tag); putting `expert` first means
          dev:mock surfaces all advanced UI without manual toggling.
        enum:
        - expert
        - advanced
        - basic
        example: expert
        type: string
    type: object
  api.SettingsResponse:
    properties:
      data:
        $ref: '#/definitions/api.SettingsData'
      success:
        example: true
        type: boolean
    type: object
...
  api.SingboxRouterSettingsData:
    properties:
      bypassExtraPorts:
        description: |-
          BypassExtraPorts is a user-defined comma-separated list of extra
          port exclusions in "PORT UDP|TCP" format (e.g. "51820 UDP, 1194 TCP").
        example: 51820 UDP
        type: string
      bypassPresets:
        description: |-
          BypassPresets lists active named port-bypass presets.
          Valid values: "l2tp", "ntp", "netbios-smb".
        example:
        - l2tp
        items:
          type: string
        type: array
      deviceMode:
        enum:
        - policy
        - all
        example: policy
        type: string
      enabled:
        example: true
        type: boolean
      policyName:
        example: awgm-router
        type: string
      refreshDailyTime:
        example: "03:00"
        type: string
      refreshIntervalHours:
        example: 24
        type: integer
      refreshMode:
        example: interval
        type: string
      snifferEnabled:
        example: true
        type: boolean
      wanAutoDetect:
        description: |-
          WANAutoDetect / WANInterface form a two-field discriminator:
            true  + ""    → sing-box auto_detect_interface
            false + "ppp0"→ sing-box default_interface=ppp0
          Other combinations are rejected by the backend validator.
          Example below shows the PINNED case as it's the more interesting
          shape to document (auto case has WANInterface omitted via omitempty
          and wanAutoDetect=true); both examples are intentionally consistent.
        example: false
        type: boolean
      wanInterface:
        example: ppp0
        type: string
    type: object
  api.SingboxRouterSettingsResponse:
    properties:
      data:
        $ref: '#/definitions/api.SingboxRouterSettingsData'
      success:
        example: true
        type: boolean
    type: object
...
  api.UpdateSettingsDTO:
    properties:
      checkEnabled:
        example: true
        type: boolean
    type: object
...
info:
  contact: {}
  description: HTTP API for the AWG Manager daemon (tunnels, routing, system).
  title: AWG Manager API
  version: "1.0"
paths:
...
  /settings/get:
    get:
      description: Returns the full Settings object (server, pingCheck, logging, dnsRoute,
        managed, apiKey, ...).
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.SettingsResponse'
        "405":
          description: Method Not Allowed
          schema:
            $ref: '#/definitions/api.APIErrorEnvelope'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/api.APIErrorEnvelope'
      security:
      - CookieAuth: []
      summary: Get settings
      tags:
      - settings
  /settings/regenerate-api-key:
    post:
      description: 'Generates a fresh UUID v4 via crypto/rand, stores it into Settings.ApiKey,
        and returns the updated Settings. The new key takes effect immediately as
        a `Authorization: Bearer <key>` substitute for the session cookie.'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.SettingsResponse'
        "405":
          description: Method Not Allowed
          schema:
            $ref: '#/definitions/api.APIErrorEnvelope'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/api.APIErrorEnvelope'
      security:
      - CookieAuth: []
      summary: Regenerate API key
      tags:
      - settings
  /settings/update:
    post:
      consumes:
      - application/json
      description: 'Persists Settings via patch semantics: any field omitted from
        the payload is preserved, including top-level bool flags. Send only the fields
        you want to change, or send the full Settings object to update everything
        atomically. ApiKey preserved when omitted (rotate via /settings/regenerate-api-key).'
      parameters:
      - description: Settings patch — any subset of fields
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/api.SettingsData'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.SettingsResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api.APIErrorEnvelope'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/api.APIErrorEnvelope'
      security:
      - CookieAuth: []
      summary: Update settings
      tags:
      - settings
...
  /singbox/router/settings:
    get:
      description: Reads the current singbox-router settings (policy mode, defaults,
        ...).
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.SingboxRouterSettingsResponse'
        "405":
          description: Method Not Allowed
          schema:
            $ref: '#/definitions/api.APIErrorEnvelope'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/api.APIErrorEnvelope'
      security:
      - CookieAuth: []
      summary: Get singbox-router settings
      tags:
      - singbox-router
    post:
      consumes:
      - application/json
      description: Persists singbox-router settings. The router is restarted only
        when fields that affect the running config change.
      parameters:
      - description: Singbox-router settings payload
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/api.SingboxRouterSettingsData'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.OkResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api.APIErrorEnvelope'
        "405":
          description: Method Not Allowed
          schema:
            $ref: '#/definitions/api.APIErrorEnvelope'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/api.APIErrorEnvelope'
      security:
      - CookieAuth: []
      summary: Update singbox-router settings
      tags:
      - singbox-router
    put:
      consumes:
      - application/json
      description: Persists singbox-router settings. The router is restarted only
        when fields that affect the running config change.
      parameters:
      - description: Singbox-router settings payload
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/api.SingboxRouterSettingsData'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.OkResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api.APIErrorEnvelope'
        "405":
          description: Method Not Allowed
          schema:
            $ref: '#/definitions/api.APIErrorEnvelope'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/api.APIErrorEnvelope'
      security:
      - CookieAuth: []
      summary: Update singbox-router settings
      tags:
      - singbox-router
...
securityDefinitions:
  CookieAuth:
    description: Session cookie set by POST /auth/login. Omit for public routes.
    in: cookie
    name: awg_session
    type: apiKey
swagger: "2.0"